Generate Cell Positivity

Generate the probability of a cell being positive given a set of simulation parameters for each file in a SpatSimObj.

Usage

GenerateCellPositivity(
  sim_object,
  k = NA,
  xmin = NA,
  xmax = NA,
  ymin = NA,
  ymax = NA,
  sdmin = 1/2,
  sdmax = 2,
  probs = c(0, 1),
  Force = FALSE,
  density_heatmap = FALSE,
  step_size = 1,
  cores = 1,
  correlation = 0
)

Arguments

sim_object

A SpatSimObj object containing the simulated data.

k

An integer specifying the number of clusters for each simulated patterns

xmin

A numeric value specifying the minimum x value for the kernel.

xmax

A numeric value specifying the maximum x value for the kernel.

ymin

A numeric value specifying the minimum y value for the kernel.

ymax

A numeric value specifying the maximum y value for the kernel.

sdmin

A numeric value specifying the minimum standard deviation for the kernel.

sdmax

A numeric value specifying the maximum standard deviation for the kernel.

probs

A numeric vector of length 2 specifying the minimum and maximum probability values for scaling the kernel values.

Force

A logical value indicating whether to force simulation parameters to be within the simulation window limits.

density_heatmap

A logical value indicating whether to compute a density heatmap for each cell.

step_size

A numeric value specifying the step size for the grid of points within the window.

cores

An integer value specifying the number of cores to use for parallel computation.

correlation

A value between -1 and 1 for how related a second or more cell type is to the first

Value

Returns the original sc.SpatialSIM object with additional generated data added to each cell object.

Details

The function generates the probability of a cell being positive given a set of simulation parameters f or each file in a sc.SpatialSIM object. It creates a kernel parameter list for k clusters in each simulated pattern and computes the probability of each point in the grid of points within the window for each cell. The function also computes a density heatmap for each cell if density_heatmap is set to TRUE.